agents_catalog/multi_agent_collaboration/00-vista-agents/cdk/requirements.txt
aws-cdk-lib>=2.201.0
constructs>=10.0.0,<11.0.0
boto3>=1.38.20
botocore>=1.38.20
PyYAML>=6.0
